HOT-2191: CMS key generation
